search: Find the closest search for an asset
From all searches that haven't been completed yet,
determine which search starts at the closest point
to the position given.

As part of the result, return the distance to and
total track length of the search, as well as a link
to get the full search data.

Adds urls for each search type to get the data for
a given search in json.