performance: Add indexes and decrease polling intervals
Add indexes for commonly filtered fields.
For asset position history this cut the get time by
an order of magnitude.

In the main map, decrease the polling rates and group
realtime data by type so the polling rate is easy to
adjust in the future.

Also, put a check around the asset update track code
so it wont perform multiple queries at the same time.