data: Get only the new asset track points
Allow the starting point (in time) to be specified for
an assets track. Update the map to make use of this
feature to limit the data transfered each time the asset
track is updated.

Also, allow switching the order of asset tracks so most
recent is at the end, rather than the start.