search: Move code from views to models
Move the code that creates searches into the model defintion
for each search (this is where the create is described, also
the create functions are db heavy)
Add functions to the search objects to calculate the
distance from a point and the total length of the search.
This replaces custom query code that was being used in views.

Also, pass the search parameters to the create functions
via objects that check the parameters.