Remove the mission_id parameter from most view functions
Most of the time, this parameter is only used by the
decorator that checks if the user is a member/admin for
the mission.
Stripping it from the parameters means less unused parameter
warnings.
In the few cases where it is still needed, mission_user.mission.pk
has the same value.