Improve asset commands
Add new commands for completing the mission and abandoning the current search.
Reply to position reporting with the current command even if the mission is complete.
Remove all assets from the mission when the mission is completed (and tell the assets this)
Set the asset command to continue when an asset joins a mission.