search: ignore inprogress searches in other missions
Sometimes a mission can be closed while searches are in progress.
If the asset is then assigned to a new mission,
it was still being given the search it was doing previously,
this isn't expected, hard to identify and could result in a fly-away.
So now we only consider searches in the mission the asset is currently
assigned to when deciding if there is an in-progress search to return.