map: Add the ability to show all (current) missions on one map
This makes it possible to merge all the searches/user data for
missions a user is in onto one map display.
The create/modify options are disabled in this view because their
usage is likely to result in unexpected results.
In the future, it might be beneficial to show which mission an item is from.