search: use the geometry version of st_azimuth
When running on debian bullseye we see an issue where the geography
version of st_azimuth doesn't return the correct value in all cases.
At least until the next debian release is available, this is the best
way to avoid this issue.

This issue manifests as creeping line ahead searches not following
the line they are created from, in some cases they will go in the opposite
direction, and in others they will deviate away.
The order/direction the line is draw has an impact on whether the
issue will occur or not.