search: remove the mission from search urls
There is no need to pass in which mission is applicable here because the
search knows which mission it is associated with.
Also, the checking on the mission was wrong, because it did check if the user
was a member of the mission, but failed to check if search was actually from
that mission, which would have allowed cross-mission search changes.
Now the mission is extracted from the search to check the user is allowed
to modify the search.