Allow users to be tracked on a mission
This is part of #317
Making individuals be treated as assets is a lot of admin work,
instead we can allow users to record their position directly into
a mission they are currently a member of.

Trails of each user can be individually selected, just like assets.
They have their own layer control to make it easier to differentiate.