usergeo: Add a common view class for viewing/deleting usergeo
Instead of needing to know the type in advance, make the usergeo
object urls not require knowning the type since they are all stored
in the same table underneath and we might want to introduce more
types in the future.
Plus, this simplifies other code that wants to link to a usergeo
object that might not know the type, e.g. linking from the search
details to the usergeo object it was created from